+++ examples/i18n/react/src/index.tsx
@@ -1,14 +1,15 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import App from "./App";
 import "./i18n";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container!);
+root.render(
     <React.StrictMode>
         <React.Suspense fallback="loading">
             <App />
         </React.Suspense>
     </React.StrictMode>,
-    document.getElementById("root"),
 );