+++ packages/antd/src/hooks/useMenu/useMenu.spec.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { renderHook } from "@testing-library/react-hooks";
+import { renderHook } from "@testing-library/react";
 
 import { MockJSONServer, TestWrapper } from "@test";