+++ packages/core/src/components/layoutWrapper/defaultLayout/index.tsx
@@ -1,5 +1,7 @@
 import React from "react";
 
-export const DefaultLayout: React.FC = ({ children }) => {
+import { LayoutProps } from "../../../interfaces";
+
+export const DefaultLayout: React.FC<LayoutProps> = ({ children }) => {
     return <div>{children}</div>;
 };