+++ packages/core/src/contexts/unsavedWarn/index.tsx
@@ -1,10 +1,12 @@
-import React, { useState } from "react";
+import React, { ReactNode, useState } from "react";
 
 import { IUnsavedWarnContext } from "./IUnsavedWarnContext";
 
 export const UnsavedWarnContext = React.createContext<IUnsavedWarnContext>({});
 
-export const UnsavedWarnContextProvider: React.FC = ({ children }) => {
+export const UnsavedWarnContextProvider: React.FC<{ children: ReactNode }> = ({
+    children,
+}) => {
     const [warnWhen, setWarnWhen] = useState(false);
 
     return (