+++ packages/core/src/hooks/data/useCreateMany.spec.tsx
@@ -1,4 +1,4 @@
-import { renderHook } from "@testing-library/react-hooks";
+import { act, renderHook, waitFor } from "@testing-library/react";
 
 import { MockJSONServer, TestWrapper } from "@test";
 
@@ -6,22 +6,19 @@ import { useCreateMany } from "./useCreateMany";
 
 describe("useCreateMany Hook", () => {
     it("should work with rest json server", async () => {
-        const { result, waitForNextUpdate, waitFor } = renderHook(
-            () => useCreateMany(),
-            {
-                wrapper: TestWrapper({
-                    dataProvider: MockJSONServer,
-                    resources: [{ name: "posts" }],
-                }),
-            },
-        );
+        const { result } = renderHook(() => useCreateMany(), {
+            wrapper: TestWrapper({
+                dataProvider: MockJSONServer,
+                resources: [{ name: "posts" }],
+            }),
+        });
 
-        result.current.mutate({ resource: "posts", values: [{ id: 1 }] });
-
-        await waitForNextUpdate();
+        await act(async () => {
+            result.current.mutate({ resource: "posts", values: [{ id: 1 }] });
+        });
 
         await waitFor(() => {
-            return result.current.isSuccess;
+            expect(result.current.isSuccess).toBeTruthy();
         });
 
         const { status, data } = result.current;
@@ -35,30 +32,27 @@ describe("useCreateMany Hook", () => {
         it("publish live event on success", async () => {
             const onPublishMock = jest.fn();
 
-            const { result, waitForNextUpdate, waitFor } = renderHook(
-                () => useCreateMany(),
-                {
-                    wrapper: TestWrapper({
-                        dataProvider: MockJSONServer,
-                        resources: [{ name: "posts" }],
-                        liveProvider: {
-                            unsubscribe: jest.fn(),
-                            subscribe: jest.fn(),
-                            publish: onPublishMock,
-                        },
-                    }),
-                },
-            );
-
-            result.current.mutate({
-                resource: "posts",
-                values: [{ id: 1 }, { id: 2 }],
+            const { result } = renderHook(() => useCreateMany(), {
+                wrapper: TestWrapper({
+                    dataProvider: MockJSONServer,
+                    resources: [{ name: "posts" }],
+                    liveProvider: {
+                        unsubscribe: jest.fn(),
+                        subscribe: jest.fn(),
+                        publish: onPublishMock,
+                    },
+                }),
             });
 
-            await waitForNextUpdate();
+            await act(async () => {
+                result.current.mutate({
+                    resource: "posts",
+                    values: [{ id: 1 }, { id: 2 }],
+                });
+            });
 
             await waitFor(() => {
-                return result.current.isSuccess;
+                expect(result.current.isSuccess).toBeTruthy();
             });
 
             expect(onPublishMock).toBeCalled();